backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4 
djoser==2.3.1